Create EID.c
During Eid, it's a tradition that each father gives his kids money which they can spend on entertainment.

Chef has N coins; let's denote the value of coin i by v i. Since today is Eid, Chef is going to give one coin to each of his two children. He wants the absolute value of the difference between the values of coins given to the two children to be as small as possible, so that he would be as fair as possible.

Help Chef by telling him the minimum possible difference between the values of the coins given to the two children. Of course, Chef cannot give the same coin to both children.